+++ README.md
@@ -82,6 +82,7 @@ Start the server
         - 로그인 ✅
         - 강의 생성  ✅
         - 리프레시 토큰 처리 ✅
+        - 강의 종료 💤
         - 실시간 코드 모니터링 💤
         - 학생 창 제어 💤
         - 일괄 코드 및 드론 실헹 제어 💤
