+++ src/features/lecture/hooks/useCreateLectureApi.ts
@@ -2,7 +2,7 @@ import { useMutation, useQuery } from "@tanstack/react-query";
 import { ICreateLectureParams } from "../../../shared/types/lecture.ts";
 import { lectureApi } from "../api.ts";
 
-export const useCreateLecture = () => {
+export const useCreateLectureApi = () => {
   return useMutation({
     mutationFn: (params: ICreateLectureParams) =>
       lectureApi.createLecture(params),
